requirements.txt
python-telegram-botbeautifulsoup4